+++ lib/features/auth/data/local_auth_service.dart
@@ -1,24 +1,2 @@
-import 'package:personal_finance/core/security/security_preferences.dart';
-
-/// Manages local auth session state and onboarding flag.
-///
-/// Backed by [SecurityPreferences] (platform secure enclave) instead of
-/// plaintext SharedPreferences, preventing bypass on rooted/jailbroken
-/// devices that can modify SharedPreferences.xml / NSUserDefaults.
-///
-/// IMPORTANT: this is a routing cache only. FirebaseAuth.currentUser is
-/// always the authoritative source — never grant data access based solely
-/// on this flag.
-class LocalAuthService {
-  Future<bool> isFirstTime() =>
-      SecurityPreferences.getOnboardingComplete().then((complete) => !complete);
-
-  Future<void> completeOnboarding() =>
-      SecurityPreferences.setOnboardingComplete();
-
-  Future<bool> isLoggedIn() => SecurityPreferences.getIsLoggedIn();
-
-  Future<void> login() => SecurityPreferences.setLoggedIn(value: true);
-
-  Future<void> logout() => SecurityPreferences.setLoggedIn(value: false);
-}
+// Movido a data/datasources/local_auth_service.dart
+export 'package:personal_finance/features/auth/data/datasources/local_auth_service.dart';
